Refactor `ResourcePool`
- Use `IResourcePool` abstract interface for public API
- Move individual pool classes to 'src', under `namespace detail`
- Concretize in 'engine.cpp'

This hides unecessary implementation details from users, and reduces includes in the public API.
