+++ lib/src/engine.cpp
@@ -1,3 +1,5 @@
+#include <detail/buffer_pool.hpp>
+#include <detail/pipeline_pool.hpp>
 #include <klib/assert.hpp>
 #include <kvf/is_positive.hpp>
 #include <kvf/util.hpp>
@@ -6,8 +8,39 @@
 
 namespace tkge
 {
+	namespace
+	{
+		class ResourcePool : public graphics::IResourcePool
+		{
+		  public:
+			explicit ResourcePool(gsl::not_null<kvf::RenderDevice*> renderDevice, vk::SampleCountFlagBits framebufferSamples)
+				: _pipelinePool(renderDevice, framebufferSamples), _bufferPool(renderDevice)
+			{
+			}
+
+			[[nodiscard]] vk::PipelineLayout PipelineLayout() const final { return _pipelinePool.PipelineLayout(); }
+
+			[[nodiscard]] vk::Pipeline GetPipeline(const graphics::Shader& shader, const graphics::PipelineFixedState& state) final
+			{
+				return _pipelinePool.GetPipeline(shader, state);
+			}
+
+			[[nodiscard]] graphics::Buffer& AllocateBuffer(const vk::BufferUsageFlags usage, const vk::DeviceSize size) final
+			{
+				return _bufferPool.Allocate(usage, size);
+			}
+
+			void NextFrame() { _bufferPool.NextFrame(); }
+
+		  private:
+			detail::PipelinePool _pipelinePool;
+			detail::BufferPool _bufferPool;
+		};
+	} // namespace
+
 	Engine::Engine(const WindowSurface& surface, const vk::SampleCountFlagBits aa)
-		: _window(CreateWindow(surface)), _renderDevice(_window.get()), _renderPass(&_renderDevice, aa), _resourcePool(&_renderDevice, aa)
+		: _window(CreateWindow(surface)), _renderDevice(_window.get()), _renderPass(&_renderDevice, aa),
+		  _resourcePool(std::make_unique<ResourcePool>(&_renderDevice, aa))
 	{
 		_renderPass.set_color_target();
 	}
@@ -19,7 +52,7 @@ namespace tkge
 	vk::CommandBuffer Engine::NextFrame()
 	{
 		_cmd = _renderDevice.next_frame();
-		_resourcePool.NextFrame();
+		static_cast<ResourcePool*>(_resourcePool.get())->NextFrame(); // NOLINT(cppcoreguidelines-pro-type-static-cast-downcast)
 		return _cmd;
 	}
 
@@ -41,7 +74,7 @@ namespace tkge
 		if (!kvf::is_positive(framebufferSize)) { return {}; }
 
 		_renderPass.clear_color = clear.to_linear();
-		return graphics::Renderer{&_renderPass, &_resourcePool, _cmd, framebufferSize};
+		return graphics::Renderer{&_renderPass, _resourcePool.get(), _cmd, framebufferSize};
 	}
 
 	void Engine::Present()
